BackEnd/requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==3.7.1
async==0.6.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
databases==0.8.0
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
fastapi==0.109.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
jose==1.0.0
jwt==1.3.1
MarkupSafe==2.1.4
numpy==1.26.3
openpyxl==3.1.2
orjson==3.9.12
pandas==2.2.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.0.3
pydantic_core==2.16.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.8.0
tzdata==2023.4
ujson==5.9.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
